Conflicts:
	changelogs/ee_changelog.txt
	system/expressionengine/libraries/Core.php
	system/expressionengine/libraries/api/Api_channel_entries.php
	system/expressionengine/views/account/ignore_list.php
	system/expressionengine/views/members/delete_confirm.php
	system/expressionengine/views/members/ip_search.php
	system/expressionengine/views/members/register.php
	system/expressionengine/views/sites/edit_form.php
	system/expressionengine/views/tools/import_from_xml.php

+++ system/expressionengine/controllers/cp/tools_communicate.php
@@ -581,8 +581,13 @@ class Tools_communicate extends CP_Controller {
 			{
 				$error = TRUE;
 			}
-
-			$this->_delete_attachments(); // Remove attachments now
+			
+			// Remove attachments only if member groups or mailing lists
+			// don't need them
+			if (empty($emails))
+			{
+				$this->_delete_attachments();
+			}
 
 			if ($error == TRUE)
 			{
@@ -1304,7 +1309,10 @@ class Tools_communicate extends CP_Controller {
 	{
 		foreach ($this->attachments as $file)
 		{
-			unlink($file);
+			if (file_exists($file))
+			{
+				unlink($file);
+			}
 		}
 	}
